SSCS-6316 - remove pdf service client dependency if not needed
