+++ src/main/java/uk/gov/hmcts/reform/sscs/service/CcdPdfService.java
@@ -1,23 +1,15 @@
 package uk.gov.hmcts.reform.sscs.service;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.*;
-import java.util.stream.Collectors;
-
+import java.util.ArrayList;
+import java.util.List;
 import lombok.extern.slf4j.Slf4j;
-import org.apache.poi.util.IOUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import uk.gov.hmcts.reform.pdf.service.client.PDFServiceClient;
-import uk.gov.hmcts.reform.sscs.ccd.domain.Correspondence;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseData;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsCaseDetails;
 import uk.gov.hmcts.reform.sscs.ccd.domain.SscsDocument;
 import uk.gov.hmcts.reform.sscs.ccd.exception.CcdException;
 import uk.gov.hmcts.reform.sscs.ccd.service.CcdService;
-import uk.gov.hmcts.reform.sscs.exception.PdfGenerationException;
-import uk.gov.hmcts.reform.sscs.idam.IdamService;
 import uk.gov.hmcts.reform.sscs.idam.IdamTokens;
 
 @Service
@@ -28,13 +20,7 @@ public class CcdPdfService {
     private PdfStoreService pdfStoreService;
 
     @Autowired
-    private PDFServiceClient pdfServiceClient;
-
-    @Autowired
     private CcdService ccdService;
-
-    @Autowired
-    private IdamService idamService;
 
     // can be removed once COR team decide what to pass into the documentType field
     public SscsCaseData mergeDocIntoCcd(String fileName, byte[] pdf, Long caseId, SscsCaseData caseData, IdamTokens idamTokens) {
@@ -47,42 +33,6 @@ public class CcdPdfService {
 
     public SscsCaseData mergeDocIntoCcd(String fileName, byte[] pdf, Long caseId, SscsCaseData caseData, IdamTokens idamTokens, String description, String documentType) {
         return updateAndMerge(fileName, pdf, caseId, caseData, idamTokens, description, documentType);
-    }
-
-    public SscsCaseData mergeCorrespondenceIntoCcd(SscsCaseData sscsCaseData, Correspondence correspondence) {
-        Map<String, Object> placeholders = new HashMap<>();
-        placeholders.put("body", correspondence.getValue().getBody());
-        placeholders.put("subject", correspondence.getValue().getSubject());
-        placeholders.put("sentOn", correspondence.getValue().getSentOn());
-        placeholders.put("from", correspondence.getValue().getFrom());
-        placeholders.put("to", correspondence.getValue().getTo());
-
-        byte[] template;
-        try {
-            template = getSentEmailTemplate();
-        } catch (IOException e) {
-            throw new PdfGenerationException("Error getting template", e);
-        }
-
-        byte[] pdf = pdfServiceClient.generateFromHtml(template, placeholders);
-        String filename = String.format("%s %s.pdf", correspondence.getValue().getEventType(), correspondence.getValue().getSentOn());
-        List<SscsDocument> pdfDocuments = pdfStoreService.store(pdf, filename, correspondence.getValue().getCorrespondenceType().name());
-        final List<Correspondence> correspondences = pdfDocuments.stream().map(doc ->
-                correspondence.toBuilder().value(correspondence.getValue().toBuilder()
-                        .documentLink(doc.getValue().getDocumentLink())
-                        .build()).build()
-        ).collect(Collectors.toList());
-
-        List<Correspondence> existingCorrespondence = sscsCaseData.getCorrespondence() == null ? new ArrayList<>() : sscsCaseData.getCorrespondence();
-        List<Correspondence> allCorrespondence = new ArrayList<>(existingCorrespondence);
-        allCorrespondence.addAll(correspondences);
-        allCorrespondence.sort(Comparator.reverseOrder());
-        sscsCaseData.setCorrespondence(allCorrespondence);
-
-        IdamTokens idamTokens = idamService.getIdamTokens();
-        SscsCaseDetails caseDetails = updateCaseInCcd(sscsCaseData, Long.parseLong(sscsCaseData.getCcdCaseId()), "uploadDocument", idamTokens, "added correspondence");
-
-        return caseDetails.getData();
     }
 
     private SscsCaseData updateAndMerge(String fileName, byte[] pdf, Long caseId, SscsCaseData caseData, IdamTokens idamTokens, String description, String documentType) {
@@ -120,11 +70,6 @@ public class CcdPdfService {
                     + caseData.getCaseReference() + "] with event [" + eventId + "]", ccdEx);
             return SscsCaseDetails.builder().build();
         }
-    }
-
-    private byte[] getSentEmailTemplate() throws IOException {
-        InputStream in = getClass().getResourceAsStream("/templates/sent_email.html");
-        return IOUtils.toByteArray(in);
     }
 
 }
